[refactor] 의미에 좀더 맞는 파일명으로 변경
