+++ test/ui/view/user/change_profile_view_test.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import '../../ink_well_button.dart';
+import '../../widget_tap_helper.dart';
 
 
 void main() {
